+++ src/components/Mypage/MyQnaListBox.tsx
@@ -2,7 +2,7 @@ import { styled } from 'styled-components';
 import { Palette } from 'styles/Palette';
 import Typo from 'styles/Typo';
 import { MyPageBoxContainer, Row } from 'assets/common';
-import { CategoryLabel } from 'components/CategoryLabel';
+import { CategoryLabel } from 'components/Category/CategoryLabel';
 
 interface qnaBoxInterface {
   children: string;
